feat: update and refactor colorScheme
